fs: compat_binfmt_elf: Add ifdef for CONFIG_ELF_CORE
The function cputime_to_compat_timeval is used by code in
binfmt_elf.c file. This file contains code using this
function that is compiled only if CONFIG_ELF_CORE option
is activated.

When this config option is not activated, build system will
obviously generate a unused-function warning, which is
stopping the build on -Wall.

Signed-off-by: Erik Castricum <git@cas-online.nl>
Change-Id: I590c942c1c79e0d97e803016056fe458651a6eca