drivers: clk: msm: make sure old_safe_freq is initialized
having old_safe_freq uninitialised leads to having random values.
The now random value is set as a safe_freq near the end of mux_div_clk_set_rate().
This newly set safe_freq is then used for comparisons against new rates that are to be set, and that
obviously fails, because the safe_freq value was random in the first place!
Intialise old_safe_freq to 0 to avoid such cases.
This was the major issue causing sony tone devices (msm8996) to have severely degraded 3d
performance!