bp: wireless: brcmfmac: Bring down bus after sending D3 INFORM
Some PCI-E BCM cards, like BCM4359, do need the bus to be up
and running to be able to process MB data, including the D3
INFORM message, otherwise the command may never reach the chip.
For this reason, end the D3_INFORM message just before bringing
down the bus and then wait for the mbdata response like normal.

While at it, make sure that the interrupts are enabled by
calling the interrupt enable function, because the mbdata
response wait_event needs an MSI interrupt to fire to get its
thread running to check the mailbox.