bp: wireless: brcmfmac: Do not try to re-register OOB interrupt
We register the OOB interrupt after firmware initialization for
cleanness purposes and to avoid an interrupt mess during init.
This gets done in the setup function, which may be called more
than once in the life of the driver for reinitialization purposes
when the chip is stuck in some bad state.

To avoid re-registration of the oob interrupt handler, add a
boolean to remember the registration state and bail out clean
if the ISR has been already set, avoiding to register it multiple
times and avoiding to create any sort of issues.