bp: wireless: brcmfmac: Force brcmf bus up when sending D3 mbdata
When we reach the suspend handler, the brcmf bus may be down
because there may have been no network activity after the last
time we resumed the chip: this is definitely not okay for some
chips that require the brcmf bus to be up to get the mbdata to
actually be sent to the chip firmware.

To address this issue, force change the state of the brcmf bus
up before sending the D3 inform mbdata to the chip.
The bus will anyway get down again as before after sending the
data.