bp: wireless: brcmfmac: Ping MAILBOX_1 on mbdata send
Writing to SBMBX in the config space would be a lot faster
than pinging the MAILBOX_1 core register, but some cards,
like BCM4359(6) have a tendency to restrict host memory
access and miss the SBMBX ping, which will produce a timeout
in the D3 suspend handler most of the times or will get the
reply back in a very large time frame (with a maximum of two
seconds).

When writing to the MAILBOX_1, we are slow, but consistently
faster than the previous way, as now it takes around 50ms.